+++ tests/test_qasm3_exporter.py
@@ -55,7 +55,7 @@ def test_to_qasm3_random_circuit(fx_bg: PCG64, jumps: int) -> None:
     depth = 5
     circuit = rand_circuit(nqubits, depth, rng=rng)
     pattern = circuit.transpile().pattern
-    pattern = pattern.infer_pauli_measurements()
+    pattern.infer_pauli_measurements()
     pattern.remove_pauli_measurements()
     pattern.minimize_space()
     _qasm3 = pattern_to_qasm3(pattern)