+++ tests/test_qasm3_exporter_to_qiskit.py
@@ -119,7 +119,7 @@ def test_to_qasm3_random_circuit(fx_bg: PCG64, jumps: int) -> None:
     depth = 5
     circuit = rand_circuit(nqubits, depth, rng=rng, use_j=True)
     pattern = circuit.transpile().pattern
-    pattern = pattern.infer_pauli_measurements()
+    pattern.infer_pauli_measurements()
     pattern.remove_pauli_measurements()
     pattern.minimize_space()