+++ tests/test_tnsim.py
@@ -330,7 +330,7 @@ class TestTN:
         pattern = circuit.transpile().pattern
         pattern.standardize()
         pattern.shift_signals()
-        pattern = pattern.infer_pauli_measurements()
+        pattern.infer_pauli_measurements()
         pattern.remove_pauli_measurements()
         state = circuit.simulate_statevector().statevec
         tn_mbqc = pattern.simulate_pattern(backend="tensornetwork", rng=fx_rng)
@@ -349,7 +349,7 @@ class TestTN:
         pattern = circuit.transpile().pattern
         pattern.standardize()
         pattern.shift_signals()
-        pattern = pattern.infer_pauli_measurements()
+        pattern.infer_pauli_measurements()
         pattern.remove_pauli_measurements()
         state = circuit.simulate_statevector().statevec
         tn_mbqc = pattern.simulate_pattern(backend="tensornetwork", graph_prep="sequential", rng=fx_rng)
@@ -398,7 +398,7 @@ class TestTN:
         pattern = circuit.transpile().pattern
         pattern.standardize()
         pattern.shift_signals()
-        pattern = pattern.infer_pauli_measurements()
+        pattern.infer_pauli_measurements()
         pattern.remove_pauli_measurements()
         state = circuit.simulate_statevector().statevec
         tn_mbqc = pattern.simulate_pattern(backend="tensornetwork", rng=fx_rng)