+++ tests/test_visualization.py
@@ -158,7 +158,7 @@ def example_hadamard() -> Pattern:
 
 def example_local_clifford() -> Pattern:
     pattern = example_hadamard()
-    pattern = pattern.infer_pauli_measurements()
+    pattern.infer_pauli_measurements()
     pattern.remove_pauli_measurements()
     return pattern
 
@@ -256,9 +256,9 @@ def test_draw_graph_reference(flow_and_not_pauli_presimulate: bool) -> Figure:
         # Pauli flow extraction from pattern is not implemented yet;
         # therefore, the pattern should not contain Pauli measurements
         # to have causal flow.
-        pattern = pattern.to_bloch()
+        pattern.blochify()
     else:
-        pattern = pattern.infer_pauli_measurements()
+        pattern.infer_pauli_measurements()
         pattern.remove_pauli_measurements()
     pattern.standardize()
     pattern.draw(